requirements.txt
streamlit
ruamel.yaml
influxdb
pandas